Consolidate all setvalue tests without modification
